+++ Changelog.md
@@ -1,3 +1,7 @@
+## 14.533.1
+* Fix for Cleave and Graze not working properly checking in some cases.
+* Compatibility bump for Foundry v14.363 and D&D5E v5.3.3
+
 ## 14.531.2.1
 * I lied. But now I did fix for errors reading `turnNumber` out of combat...